+++ onegov/ballot/__init__.py
@@ -5,6 +5,7 @@ from onegov.ballot.models import Ballot
 from onegov.ballot.models import BallotResult
 from onegov.ballot.models import Candidate
 from onegov.ballot.models import CandidateResult
+from onegov.ballot.models import ComplexVote
 from onegov.ballot.models import Election
 from onegov.ballot.models import ElectionResult
 from onegov.ballot.models import List
@@ -12,6 +13,7 @@ from onegov.ballot.models import ListConnection
 from onegov.ballot.models import ListResult
 from onegov.ballot.models import PanachageResult
 from onegov.ballot.models import PartyResult
+from onegov.ballot.models import ProporzElection
 from onegov.ballot.models import Vote
 
 __all__ = [
@@ -20,6 +22,7 @@ __all__ = [
     'BallotResult',
     'Candidate',
     'CandidateResult',
+    'ComplexVote',
     'Election',
     'ElectionCollection',
     'ElectionResult',
@@ -28,6 +31,7 @@ __all__ = [
     'ListResult',
     'PanachageResult',
     'PartyResult',
+    'ProporzElection',
     'Vote',
-    'VoteCollection'
+    'VoteCollection',
 ]