+++ onegov/ballot/models/election/__init__.py
@@ -7,6 +7,7 @@ from onegov.ballot.models.election.list_connection import ListConnection
 from onegov.ballot.models.election.list_result import ListResult
 from onegov.ballot.models.election.panachage_result import PanachageResult
 from onegov.ballot.models.election.party_result import PartyResult
+from onegov.ballot.models.election.proporz_election import ProporzElection
 
 
 __all__ = [
@@ -19,4 +20,5 @@ __all__ = [
     'ListResult',
     'PanachageResult',
     'PartyResult',
+    'ProporzElection',
 ]